Annotate luajr.lua and have SEXP_length() return double

+++ src/lua_api.cpp
@@ -197,7 +197,7 @@ extern "C" void SetPtr(void** ptr, void* val)
     *ptr = val;
 }
 
-extern "C" int SEXP_length(SEXP s)
+extern "C" double SEXP_length(SEXP s)
 {
     return Rf_length(s); // TODO XLENGTH_EX(s) ?
 }
